Load config before the app starts.
Inline JavaScript will block page parsing and will run at once,
starting to fetch the config.json, before the app bundle is fetched.
TODO: test if caching effects could cause the app to start running before the config is loaded.

Alternatively we could deliver the configuration properties as a JavaScript file,
containing also the code to assign it to the global variable. This would block until everything
is in place, but it would inject potentially arbitrary code via the configuration mechanism.